requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.101.1
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
pydantic==2.2.1
pydantic_core==2.6.1
pyngrok==6.0.0
pyTelegramBotAPI==4.13.0
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
